Fix CMS Preview Error
Error
There's been an error - please open an issue on GitHub.

Opening an issue pre-populates it with the error message and debugging data.
Please verify the information is correct and remove sensitive data if exists.

Details
RangeError: Use `d` instead of `D` for formatting days of the month; see: https://git.io/fxCyr